Migrate essential data handling to new class `AppDataHandler.java`
