+++ src/main/java/TF2ClassWarsStatTracker/gui/tracking/RecordWinButtonHandler.java
@@ -1,8 +1,8 @@
 package TF2ClassWarsStatTracker.gui.tracking;
 
+import TF2ClassWarsStatTracker.AppDataHandler;
 import TF2ClassWarsStatTracker.Start;
 import TF2ClassWarsStatTracker.exceptions.GameMapNotFoundException;
-import TF2ClassWarsStatTracker.game.GameMap;
 import TF2ClassWarsStatTracker.util.FileHandler;
 import TF2ClassWarsStatTracker.util.Print;
 
@@ -20,10 +20,10 @@ public class RecordWinButtonHandler implements ActionListener {
     @Override
     public void actionPerformed(ActionEvent e) {
         try {
-            GameMap.incrementWins(Tracking.getSelectedMap(), Tracking.getSelectedGameMode(),
+            AppDataHandler.incrementWins(Tracking.getSelectedMap(), Tracking.getSelectedGameMode(),
                     Tracking.getSelectedBluMercenary(), Tracking.getSelectedRedMercenary(),
                     team);
-            FileHandler.writeToJSONFile(GameMap.getMaps(), "res/maps.json");
+            FileHandler.writeToJSONFile(AppDataHandler.getMaps(), "res/maps.json");
         } catch (GameMapNotFoundException | IndexOutOfBoundsException ex) {
             JOptionPane.showMessageDialog(Start.getTrackerScreen(), ex.getMessage(), "Error", JOptionPane.ERROR_MESSAGE);
             Print.error(ex.getMessage());
